[TA2655] refactor(provisioner):remove non cas code
Upto 0.6 release the provisioner used different(than 0.7) structs and methods
to talk to m-apiserver.
Now the approach has changed and cas volume structs and methods are used,
thus the unused code is deleted and common function has been moved to cas
specific files.
